requirements-test.txt
# Core Django and DRF
Django==4.0.4
djangorestframework==3.13.1

# Testing
pytest==7.3.1
pytest-django==4.5.2
pytest-cov==4.0.0
pytest-mock==3.10.0

# API Testing
requests==2.31.0
responses==0.23.1

# Code quality
coverage==7.2.3
